SW-6998 Support fetching simple planting site history (#3127)
If you tried to fetch the history of a planting site that started out as a simple
planting site (even if it later turned into a detailed one) it would fail because
there's no "grid origin" value for simple planting sites and the history fetching
code was expecting one.

Update the fetching code to make grid origin values optional.

Grid origin isn't exposed in the history fetching API, so no client changes
should be needed.