Use AWS JDBC wrapper for RDS IAM authentication (#3137)
To authenticate to the RDS databases using temporary credentials, we previously
used a custom DataSource implementation. This doesn't work in Spring Boot 3.5 for
some reason, but luckily there's a better alternative: a JDBC driver from Amazon
that wraps the underlying Postgres driver and takes care of the IAM authentication
automatically.

Add the driver and use it for running tests.

The driver requires some additional configuration to tell it to enable IAM.
Introduce a new YAML config file for the "aws" Spring profile, which is used in
staging and prod, to set that property and to tell Spring to use the AWS driver.