SW-7016 Optionally return plot-level "Unknown" results
By default, when we're returning observation results, we don't include
"Unknown" species in the list of species-level results. But we need to be able
to include them in exported CSV files. Add an optional parameter to the
endpoint that fetches results for a single observation to allow the client to
say that it wants unknown species included.

This only affects the monitoring-plot-level species lists; the zone- and
site-level species are filtered in `ObservationResultsStore`. Including unknown
species at the site and zone levels breaks some of the mortality rate
calculations.  We only need the unknown species at the plot level in order to
generate correct CSV files, so there's no need to change the underlying
fetching logic here.