Make JobRunr recovery thread wait for initialization (#3140)
We have a background thread that restarts JobRunr if the database connection goes
away temporarily. Currently, during application startup, it sometimes sees that
JobRunr isn't started yet and tries to recover. This is harmless, but it spits
out a warning log message which adds unnecessary noise after a restart.

Change it to wait for JobRunr to start up initially.