SW-7000 Add basic GeoServer client
Add a client class to send requests to GeoServer with optional authentication.

Some GeoServer API endpoints can only return XML responses, not JSON ones; add the
Jackson XML serialization library and an example usage in the form of a very
stripped-down version of the endpoint that fetches the server's capabilities.