SW-7034 Remove observations that can't be started (#3155)
If a planting zone is too small for the required number of permanent and temporary
plots for an observation, we currently log an error and try again later, which
means there's a neverending trickle of error messages in the logs as we repeatedly
try and fail to start the same observations. Worse, the observations still show up
for users, listed as "upcoming" even when their scheduled start dates are past.

Since we will never be able to start an observation of a site that's too small,
make the server delete observations when there isn't room for the required number
of monitoring plots.