SW-7059 Fix divide-by-zero in observation summary (#3166)
If a monitoring plot only had one observation where we calculated a mortality
rate, we were trying to divide by zero when calculating the mortality rate's
standard deviation. The standard deviation of a single sample is zero by
definition, so we can skip the rest of the standard deviation calculation in that
case.