SW-6808 Implemented CRUD for disclaimers (#3158)
This PR implements functions to create, read, accept and delete disclaimers. 

Read and accept current disclaimer are exposed via external facing APIs.

Creating, deleting, and listing disclaimers and user acceptance records will be functions that are exposed through the admin UI, which will be part of the PR that follows.