SW-7037 Recalculate site populations from admin UI
In cases where nursery batches have been deleted, there can be a discrepancy
between the outplanting withdrawal totals from the withdrawal log and the reported
plant populations that are tracked in `planting_site_populations` and its sibling
tables.

We're figuring out a more robust plan for this going forward, but until that's
in place, we need to be able to clear up the discrepancies in the numbers.

Add an admin UI function to reset the populations to the totals from the
withdrawal log.