SW-7083 Allow native access in Docker image (#3184)
The Netty library uses a native-code component on Linux, which it needs to load
using JNI. In Java 24, this causes the JVM to output a warning that access to
native code will be denied by default in a future version. Add the command-line
argument to explicitly enable native code access.

This will stop the warning in Java 24, and will stop the application from breaking
in whatever future Java version changes the default behavior.