SW-7057 Set quantities to 0 on batch delete (#3186)
When the user deletes a batch that has had withdrawals, keep it around but set its
remaining quantities to zero. This will remove it from the list of active batches
in the web app, while preserving its entries in the withdrawal log.

The reasoning is that users are using the delete function just to clear irrelevant
batches out of their inventory, not to forget the fact that the batches ever
existed.

However, for batches that have had no withdrawals, we continue to delete them
outright. This will have no effect on the nursery's withdrawal log, but will
remove batches that were entered by mistake and never used for anything.