SW-7087 Support stdout Logstash logging (#3199)
Currently, in production and staging, we run with the `jsonlog` Spring profile
which causes the server to write Logstash (JSON) formatted logs to logfiles on
the host filesystem. The Datadog agent then forwards the logs to Datadog.

When we run in ECS, we'll want to let CloudWatch collect logs and forward them
to Datadog. But we'll still want them in Logstash format so we can log structured
information.

Update the logging configuration to support a new Spring profile `logstash` which
uses Logstash format but writes the logs to stdout instead of to files. Since the
access and application logs will both go to stdout, this adds a new `log_type`
custom field we can filter on as needed.