Support disabling JobRunr background thread (#3201)
When we launch a server instance (say, in a dev environment) that connects to the
staging database rather than a local database, we don't want that server instance
to do things like send out scheduled notifications. JobRunr has a way to disable
the background job server thread, but currently that causes terraware-server to
fail to start up because the Spring bean for our recovery thread depends on the
bean for the JobRunr background job server thread.

Add a Spring autoconfigure rule so our recovery thread bean only gets created if the
JobRunr one exists.