SW-7117 Allow retrieval of multiple published funder project details (#3204)
Previously the endpoint only allowed for one projectId to be retrieved at a time. Update the endpoint to receive a set of projectIds. Return this in a different property so that the FE doesn't break. 
Eventually the original code returning only one project's details should be removed once the FE is only using the list of project details returned.