SW-6996 Insert new species when recalculating populations
The admin UI function to recalculate planting site populations wasn't handling the
case where a subzone had plantings of a given species but the subzone/species pair
wasn't present in `planting_subzone_populations` at all. This wasn't a problem
with the original incorrect totals the recalculation function was designed to fix,
but is an issue when we're using it after manually shuffling some plantings
between two planting sites.