SW-7119 Add support for ECS deployment
For prod and staging deploys, if the repo has variables whose names are the
tier name followed by `_ECS_CLUSTER` and `_ECS_SERVICE` (for example,
`STAGING_ECS_CLUSTER`), deploy to that ECS cluster and service.

For now, we continue to deploy to the EC2 instances even if we also deploy to
ECS. This will let us test ECS deployment without disrupting the existing
prod/staging environments.