Drop batch_summaries view
We had a view of the batches table that included a couple computed fields; the
view was used in the search API because earlier revisions of the search API
required that search fields map to table (or view) columns rather than to
arbitrary expressions. That restriction has been gone for a while, and the view
adds another bit of complexity when we want to change the batches table.

Drop the view and update the search code to use the batches table directly.