SW-7102 Comment on PR when staging deploy finishes
To make it easier to see when changes are ready for QA on staging, post a comment
on the PR when a deployment from the `main` branch finishes. This relies on the
fact that we use squash merges, so there'll always be a suffix on the commit title
with the PR number.