SW-7172 Make null ids greater than non-null for same-date withdrawal sorting (#3229)
When creating a nursery transfer withdrawal for a date that already had a withdrawal, the server does not use the most recent withdrawal to check if it is a Withdraw-All operation. This throws the all-too-familiar server error: "Withdrawal quantity can't be more than remaining quantity".

Change the sorting to make null IDs greater than non-null IDs for same-date withdrawal comparison.