SW-7138 Add Permission Grants for Read Only Users (#3230)
As a part of the effort to expand the functionality of the Terraformation Contact role in Terraware, we decided we are also going to make it such that all users with Global Roles have permission to read all accelerator project data.

Update the permissions of the Read Only role so they can read all accelerator related projects and orgs.