SW-7164 Create FetchAll Published Projects Endpoint (#3228)
Product wants Funding Entities to only be able to select projects that have been published. 

Add an endpoint to retrieve all published projects.

For now it only returns `dealName` and `projectId` for each since that is all that is required at the moment. Later this endpoint could fetch all details for each if needed.