Disable H2 console (#3252)
By default, Spring Boot supports an admin console for the embedded H2 database
engine. We include H2 as a dev-only dependency so we can use it as a dummy
data source for the "apidoc" profile, but that means it's also present in dev
environments.

Spring detects its presence and sets up the console. The console isn't accessible
because there's no access control rule for it in our security configuration and
we have a deny-by-default security policy, so it's doing nothing but eating memory
and wasting time at application start. Turn it off.