Allow table names as search prefixes (#3253)
Previously, the prefix in a search API call had to be specified as a sublist path
starting from the `organizations` search table. This could be annoying when
searching against tables several steps removed from organizations.

Update the search API such that you can specify either a sublist path or the name
of a search table as defined in the `SearchTables` class.

We can update clients to use table names instead of paths, after which we can get
rid of path-based prefixes.