SW-7192 Change criteria in search service to be a map of prefix to SearchNode (#3248)
This is some pre-work for making sublist searching work in the search api, which has been accomplished in this PR stack by allowing `criteria` to be a map of prefix to SearchNode, and then applying a filter condition when adding each select field (if relevant) instead of only at the end of query building. See downstream PRs for more details.