SW-7192 Add sublist criteria support to search api (#3251)
Update NestedQueryBuilder to support sublist querying in the search api. This is accomplished by using the previously-passed `criteria` to add a search condition to a sublist's multisetQuery.

Also update `SeachController` to add the necessary property to the payload for sublist searching. 

(this should all be discussed with Steve to ensure there are no other potential issues with adding this support)

(if this approach is confirmed to be the way we want to go, better documentation is needed on NestedQueryBuilder since the top large comment says that sublist filtering is not currently supported - lines 266-285)