SW-7160 Add simple generic email sender to admin UI (#3285)
The Nursery Phase Terminology updates need a way to notify all users by email. Instead of building a one-time usage email, there was discussion around building a reusable general purpose emailer.

Add a simple generic email sender to admin UI. 

It supports html and will get rendered properly in the html version of the email, but not currently in the text version of it (can change this if preferred).

This currently does no scrubbing of the content as it assumes full trust in our super admins. 

A few things that can be updated if preferred:
- change from the emailBody being a requestParam to part of the body
- add descriptions to the new email fields
- change header/title from `Send Test Email` to `Send Email`
- remove html characters for text version of email

WYSIWYG editor in next PR.