Make notReadyQuantityWithdrawn optional in BatchWithdrawalPayload (#3308)
`BatchWithdrawalPayload` has a getter for `notReadyQuantityWithdrawn` to support backwards compatibility in response payloads (and `JsonAlias` for request payloads). Unfortunately, the `generated-schema.ts` file doesn't know that the getter is only going to be on the response, and the property is not required for the request.

Make the property optional so that `generated-schema.ts` correctly shows it as optional, allowing the renaming of the property to get completed.

(Note that this temporary getter will need to stay in place for backwards compatibility for a few months in order for mobile to update and get all users on the latest version.)