SW-7265 Add Admin UI to Delete Org Users (#3307)
We've never had a way to delete users from organizations that they don't belong to. Now that orgs can have multiple TF Contacts, these can start to pile up in an org because adding a new one no longer replaces the old one.
Add a section to the Admin UI (and corresponding endpoint) to delete users from an organization, only for Super Admins.

Currently this does not allow deleting the user if it's the last user in an org, but this may be added if product prefers.