SW-7263 Include Used Up in accession summary stats (#3311)
`GET /api/v1/seedbank/summary` previously only returned accession counts for
"active" statuses, but we want to start showing the total number of Used Up
accessions for a facility in the web app. Remove the filter on statuses.

We were logging the SQL and the query execution time for the counts-by-status
query; remove that logging since we can get that data from the database server
logs.