SW-7289 Plot boundary calculation tool
We've now had a couple cases where we need to manually add a monitoring plot to a
planting site based on the location of a previously-existing plot. Eventually
we'll want to automate that process, but for now, add an admin UI page with
a tool for one of the steps: figuring out the geometry of the new plot.