Update Internal Users api to combine calls (#3319)
As part of FE implementation, it was realized that it would be easier to update internal users with a single endpoint instead of one for adding a single one and removing a single one (especially since updating the role for a user required first calling the endpoint to remove it then calling the endpoint to add it with the new role).
Update the api to accept the full list of users and roles that will be Project Internal Users. Update the service to parse out existing users vs requested users in order to determine which to delete and which to add. For users that need to be updated, add them to both lists.