SW-7088 Automatic translation into French and Spanish (#3335)
Introduce our new LLM Autotranslate tool to the dev environment. It uses
the OpenAI API to automatically generate French and Spanish translations
for strings. README.md is updated with instructions for running it.

It detects string edits using hashcodes; these are stored in the comment
fields of the French and Spanish strings properties files.

This initial change adds hashcodes to the existing translations from
Phrase so they're grandfathered in.

To prevent us from forgetting to generate translations, a new unit test
checks whether translations have been generated for all the strings in
the files covered by autotranslate.