Implement InputStreamCopier to double-process streams
For the upcoming changes to support photos and videos in activity log entries, we
are going to need to store uploaded files in S3 and also pull metadata from them
(GPS location, etc.) We get an `InputStream` with the uploaded data, but we can
only read it once unless we want to load the whole thing into memory, impractical
with large video file uploads. We could also read the file back from S3 after
saving it, but that would incur extra transfer fees and would take extra time.

Add a new utility class that duplicates an `InputStream` incrementally, allowing
it to be read by multiple consumers at once. The source stream is incrementally
loaded into a small temporary buffer on demand, with backpressure to prevent fast
readers from outpacing slow ones by more than the size of that buffer.

We'll be able to use this to feed one copy of the stream to the file store and
another copy (or copies) to Apache Tika for metadata extraction.