Fix race in InputStreamCopierTest (#3368)
One of the tests had a race condition that caused it to fail occasionally.

What the test is verifying is difficult to test 100% deterministically (it's
kind of the halting problem) but we can make it less susceptible to races using
explicit signaling, and make it pass rather than fail if it's running on an
extremely loaded host such that its sleep time isn't long enough to trigger the
condition it's testing for.