SW-7367 Allow temporary read access to files
To allow a video streaming provider to convert user-uploaded videos to the correct
formats for streaming, the provider needs to be able to access the original file,
and of course it won't know how to authenticate to Terraware.

Add an endpoint that works similarly to features like S3's pre-signed URLs: we can
generate a URL that includes a token that allows unauthenticated access to a file
for a certain amount of time. For our purposes, there's no need for anything as
complex as signed keys; it's sufficient to just generate a hard-enough-to-guess
token and check it against our database.

The admin UI has a new page that super-admins can use to generate tokens for
testing.