Speed up ObservationResultsStoreTest
`ObservationResultsStoreTest` is one of the slower test classes in our test
suite.

Some of the observation scenarios involve recording a lot of plants. Profiling
showed that it was spending a lot of time in `ObservationStore.completePlot`,
specifically the call to `RecordedPlantsDao.insert` which was inserting the
plants rows one at a time because it needed to retrieve their
database-generated IDs.

We don't care about the IDs in that context, so we can do a batch insert
operation instead.

In rough benchmarking, this appears to speed the test up by around 30%, still
not lightning-fast but noticeably quicker.

Completing a monitoring plot will be faster for users too, but the data volumes
are low enough that it won't be noticeable.