SW-7383 Do not include survival rate when not all sub areas have it (#3403)
The PRD says this:
> If all of the observed subzones in a zone have a calculated survival rate then Terraware will calculate survival rate in addition to mortality rate for that zone.
> If all of the observed subzones in a planting site have a calculated survival rate then Terraware will calculate survival rate in addition to mortality rate for that site.

But after discussion we just care about not displaying it, and don't want the FE doing the work for determining if it should be displayed or not. 
Update the Results model to only calculate zone and site survival rate (and stddev) if their observed sub-areas have a survival rate.