Add assertSetEquals to compare sets in tests
We use sets to compare unordered collections of items in a lot of our tests.
The default JUnit `assertEquals` will correctly detect when two sets aren't
equal, but the assertion failure message is hard to read, especially when
the two sets have the same items but in a different order.

Add a helper function to compare sets and produce a more human-friendly failure
message if they aren't equal. The items are sorted by their string representations
and are separated by newlines.