SW-7428 Fix error when deleting a species density that was recently added (#3452)
When a species' density was recently added, the event has the following properties:
```
previousDensity = null,
newDensity = X,
```
When the same species is then deleted before an email is sent, it throws an error because the new event has these properties:
```
previousDensity = X,
newDensity = null,
```
resulting in a `combine` operation that sets previous and new Density to `null` which throws an error. 
When `previousDensity` and `newDensity` are equal, remove the species from the list of changed species in the `combine` method.