Properly set memory limit for test suite (#3458)
Previously, we were setting the memory limit for the Gradle worker process that
runs the test suite by adding it to Gradle's JVM arguments. That no longer works;
instead we need to include it in the configuration for the test-running task.
Otherwise Gradle will use the default limit of 512MB, which isn't enough to
run our tests in parallel without lots of memory pressure that causes frequent
GC pauses.

In a local dev environment, this change cuts test suite execution time in half.