SW-7472 Generalize two-stage thumbnail generation
For video files, we use a two-step process for generating thumbnails: first
generate a full-sized JPEG still image from the video, then scale that image to
the requested size.

We'll want to use a similar process for image file formats that aren't supported
by the Java ImageIO library, which we use for thumbnail generation.

Update the code so that the two-step process isn't specific to videos, and make
the one-step process work for all image formats we can handle natively.