SW-7472 Return full-sized images as JPEG (#3474)
If the user requests a full-sized version of an image, and the original image file
is not in JPEG or PNG format, return the full-sized JPEG "thumbnail" instead of
the original.

This will allow non-Safari users to view full-sized versions of HEIC images, since
HEIC is not supported by other browsers.