Assert that t0 data can only be assigned to permanent plots (#3480)
We discovered that any plot that had ever been permanent in an observation were being included in the Survival Rate settings page. This was causing the FE to assign t0 data to permanent plots which isn't supposed to be allowed. The data was being rolled back when the `T0PlotDataAssignedEvent` occurred and the survival rates were being recalculated. Instead of relying on a different method to assert permanence, add a permanence check to the `assign` methods.