SW-7514 Fix issue where sublist filtering depends on order of search fields (#3488)
Sublist Filtering was only working if there were no flattened fields in the sublists being filtered (or if a direct property on the sublist was being queried above).
Example:
```json
"prefix": "projects",
"fields": [
  "id",
  "internalUsers.user_firstName",
],
"filters": [{
  "prefix": "internalUsers",
  "search": {
    "operation": "field",
    "field": "role",
    "values": ["Project Lead"]
  }
}]
```
This would only work if the field `internalUsers.role` was added _above_ `internalUsers.user_firstName` because then it would get processed first and the sublist query would get built with the correct criteria. 

The problem was that `fieldPath.prefix` for `internalUsers.user_firstName` is `internalUsers.user_` which causes `criteria[fieldPath.prefix]` to be null when adding the `sublistQuery` object to `sublistQueryBuilders`. 

Change this to `criteria[prefix.relativeSublistPrefix(sublistName)]` which correctly returns `internalUsers`.

Uncomment related sublist-filtering test, and delete duplicate test.