Allow ReadOnly users to access project_internal_users in search requests (#3486)
The Matrix View now uses the search api to get project leads for projects, but users are currently only visible if they are in the same org as the current user (or if they're a `project_internal_user` and the project is in the same org as the current user. 

Update the `UsersTable` to allow ReadOnly+ users to see `project_internal_users`, even if not in the same org.