SW-7506 Return publication status of activities
Add a `publishedTime` field to the end-user activity payload.  If it's
non-null, the activity has been published. The web app can use this to decide
whether to show edit and delete controls to end users.

Add `publishedBy` and `publishedTime` fields to the admin activity payload.
These can also be used to show publication status, but also, if `publishedTime`
is non-null but is less than `modifiedTime`, it means the activity has
unpublished changes. The console can show a message to that effect.