SW-7506 Non-admins cannot modify published activities (#3495)
Don't allow end users to update or delete published activities.

This adds a feature to `TestEventPublisher` to allow tests to register event
listeners; it's needed here because without code to handle the event that's
fired during activity deletion, deleting an activity that has a published
version would fail due to a foreign key constraint.