SW-7505 Methods to read published activity media files (#3499)
Implement the internal logic for reading published activities' media files.

This differs from the similar logic for activity media files in that it requires
a different permission (which means it'll be available to funder users) and
doesn't provide a way to read original raw files, since they might contain
sensitive information.