SW-7532 Store and fetch persistent events
Create the `event_log` table with indexes on various potential IDs that events
will have. A new database function makes it easy to index additional ID fields
as needed.

`EventLogStore` provides a type-safe API for writing new events to the event log
and for querying them based on ID fields and event types, with support for
migrating events to newer versions on the fly.