SW-7532 Add persistent event interfaces (#3512)
In preparation for keeping a persistent event log, add two interfaces that event
classes will implement. `PersistentEvent` is a marker interface, and
`UpgradableEvent` defines a function that converts an older version of an event to
a newer version.

Event classes must have version number suffixes consisting of the letter `V`
followed by a positive integer. This is enforced by a new test. The test will also
catch backward-incompatible changes to existing event classes as well as some
other possible structural issues with events.