SW-7560 Log persistent events to console/Datadog
When a persistent event is published, in addition to storing it in the database,
log a message about it to the console (in dev environments) or Datadog (in staging
and production). The log message may be customized on a per-event basis.

The Datadog log message includes the event payload in JSON form, which will allow
ad-hoc searches on specific event fields.