SW-7562 Reject invalid requests for file access
There's a bot that sometimes tries fetching all the endpoints in our OpenAPI
schema. Most of them require authentication, so the requests fail. But the file
access endpoint, since it's used by Mux to fetch files for processing, needs to
be public. When the bot hits that endpoint, we try to look up the literal string
`{token}` as a token, and of course there's no such token, so we log an error
message. The error is useless noise.

Add a minimum size constraint to the token. Actual tokens are UUIDs and will
always be above the minimum size, but the string `{token}` is below the minimum.
This causes the bogus requests to be rejected with HTTP 400 by Spring, and no
error message is logged.