Remove "jsonlog" Spring profile
When we were running terraware-server on EC2 instances with the Datadog agent, we
needed it to write its JSON-formatted logs to files that the agent would monitor.
But with our ECS deployment, we no longer need that, and it's not useful in dev
environments either. Remove support for it from the logging configuration.