Prevent double CI runs for Renovate PRs (#3524)
The GitHub Actions workflow had a branch filter that was supposed to stop Renovate
PRs from running the workflow on both the initial `push` event and then on the
subsequent `pull_request` event, but it didn't work because the `branches-exclude`
filter matches against the base branch, not the PR's branch.

Replace it with a conditional in the job definition.