SW-7559 Search event log in admin UI (#3526)
Add an admin UI page that super-admin users can use to search the event log.

It shows a list of all known event classes and, depending on which ones are
selected, a filtered list of ID properties they share in common. In search
results, IDs can be clicked to set up a search for events related to that ID.