SW-7565 Fix site survival rate showing when no t0 data has been set (#3529)
The site survival rate was showing a value in the plants dashboard even if no t0 data had been set for the site. The query for the species was incorrectly retrieving t0 data for all planting sites instead of just the requested. Only retrieve t0 data for the requested planting site(s).