SW-7569 Move EXIF extraction to FileService
Move the "extract EXIF metadata from a file and store it" logic from
ActivityMediaService to FileService so it can easily be reused for file uploads in
other parts of Terraware.

With this change, we'll start recording captured dates and geolocations for photos
uploaded to all parts of the app, though those values aren't accessible via the
API in any new places yet.