SW-7569 Move geolocation, capture time to files table (#3530)
We're going to be pulling GPS coordinates from photo and video files in additional
places in the app soon. Currently we only extract them from files that are
attached to activity log entries, and we store them in `activity_media_files`.
We also store GPS locations for observation photos, though those GPS coordinates
are sent separately in the photo upload API request rather than extracted from the
photo files.

Move the geolocation and captured date columns to the files table so they'll be
stored consistently across the app. "Captured date" turns into "Captured local
time" because we may want the time of day rather than just the date; the word
"local" is to clarify that the timestamp is not necessarily in UTC. But in this
initial change, we still only extract the date part of the EXIF timestamp for
activity media files.

For observation photos, the GPS coordinates continue to come from the API request
payload for now, such that there's no client-visible behavior change.