SW-7570 Add endpoint for checking if all site t0 data has been set (#3531)
To determine whether all t0 data has been set, the FE currently has to query 3 endpoints and combine all the data, and then the data is never used again on the page.

Add an endpoint to do this determination on the BE since this is used on pages that only need to show an indication for whether or not all t0 data has been set.