SW-7581 Remove image file validation
For most of the photo upload endpoints, we've been validating that the uploaded
files can be loaded using the Java ImageIO class and rejecting them if not.
That won't work for HEIC images, which we can't read directly, or for video files.
Remove the validation since it won't be useful protection against malformed files
once we start supporting those formats in more places in the app.