Render all MJML files in one Node command
Update the `renderMjml` Gradle task to run a single command to render all the MJML
files rather than running a separate command for each file. The vast majority of
the execution time for that task is from firing up Node and loading all the MJML
modules, so this change makes the task take only slightly longer to render all the
files than it used to take to render each one.

The wrapper script we use to load our plugins is no longer a passthrough to the
MJML command-line interface; it's now specialized for rendering a batch of files.