Render all MJML files in one Node command (#3537)
Update the `renderMjml` Gradle task to run a single command to render all the MJML
files rather than running a separate command for each file. The vast majority of
the execution time for that task is from firing up Node and loading all the MJML
modules, so this change makes the task take only slightly longer to render all the
files than it used to take to render each one.

The wrapper script we use to load our plugins is no longer a passthrough to the
MJML command-line interface; it's now specialized for rendering a batch of files.

Since we're rendering MJML using the MJML library's API rather than its 
command-line tool, we need to minify the generated HTML ourselves (minification 
is only supported by MJML's command-line tool). Configure the minifier with the
same options the MJML CLI uses, but add a custom option to ignore Freemarker
template tags like `<#if>`. With that change, we no longer need to wrap
Freemarker tags in `<!-- htmlmin:ignore -->` to avoid HTML parser errors.